Add files via upload
The Play Store apps data has enormous potential to drive app-making businesses to success. However, many apps are being developed every single day and only a few of them become profitable. It is important for developers to be able to predict the success of their app and incorporate features which makes an app successful. We can collect app data and user ratings from the app stores and use it to extract insightful information. A machine learning model can be used to predict rating for a given appThe Play Store apps data has enormous potential to drive app-making businesses to success. However, many apps are being developed every single day and only a few of them become profitable. It is important for developers to be able to predict the success of their app and incorporate features which makes an app successful. We can collect app data and user ratings from the app stores and use it to extract insightful information. A machine learning model can be used to predict rating for a given app
Skills: Supervised Learning. (Decision Trees · ensemble methods